requirements.txt
aiodns==3.0.0
aiohttp==3.8.1
aiohttp-socks==0.7.1
aiosignal==1.2.0
alice-blue==1.0.4
aniso8601==9.0.1
anyio==3.6.1
app-store-scraper==0.3.5
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.4.1
asn1crypto==0.24.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
Automat==20.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==4.1.0
Brlapi==0.6.6
bs4==0.0.1
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
cmake==3.25.2
command-not-found==0.3
constantly==15.1.0
contextlib2==21.6.0
contextvars==2.4
cryptography==37.0.4
cssselect==1.1.0
cupshelpers==1.0
cycler==0.11.0
dataclasses==0.8
dateparser==1.1.1
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro-info===0.18ubuntu0.18.04.1
Django==3.2.16
django-dynamic-admin-forms==0.1.8
djangorestframework==3.14.0
djongo==1.3.6
dlib==19.24.0
dnspython==2.2.1
docutils==0.14
elastic-transport==8.1.2
elasticsearch==8.3.3
entrypoints==0.4
face-recognition==1.3.0
face-recognition-models==0.3.0
fake-useragent==0.1.11
fastapi==0.78.0
filelock==3.4.1
Flask==2.0.3
Flask-Admin==1.6.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flask-swagger-ui==4.11.1
Flask-WTF==1.0.1
frappe-bench==5.12.1
frozenlist==1.2.0
geographiclib==1.52
geopy==2.2.0
gitdb2==2.0.6
GitPython==2.1.15
google-play-scraper==1.1.0
googletransx==2.4.2
h11==0.13.0
honcho==1.1.0
httplib2==0.9.2
hyperlink==21.0.0
idna==2.10
idna-ssl==1.1.0
immutables==0.18
importlib-metadata==4.8.3
importlib-resources==5.4.0
incremental==21.3.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
itemadapter==0.7.0
itemloaders==1.0.4
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
keyring==10.6.0
keyrings.alt==3.0
Kivy==2.1.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
linkedin-scraper==2.9.1
louis==3.5.0
lxml==4.8.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==2.0.1
matplotlib==3.3.4
MechanicalSoup==1.1.0
meld==3.18.0
mistune==0.8.4
multidict==5.2.0
multitasking==0.0.10
mysql-connector==2.2.9
mysqlclient==2.1.1
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.6
netifaces==0.10.4
notebook==6.4.10
numpy==1.19.5
oauth==1.0.1
olefile==0.45.1
packaging==21.3
pafy==0.5.2
paho-mqtt==1.6.1
pandas==1.1.5
pandas-datareader==0.10.0
pandocfilters==1.5.0
parse==1.19.0
parsel==1.6.0
parso==0.7.1
pexpect==4.2.1
pickleshare==0.7.5
Pillow==8.4.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
Protego==0.2.1
protlib==1.4
protobuf==3.0.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pycairo==1.16.2
pycares==4.2.2
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
pydantic==1.9.1
PyDispatcher==2.0.5
pyee==8.2.2
Pygments==2.14.0
PyGObject==3.26.1
pymacaroons==0.13.0
pymongo==4.1.1
PyNaCl==1.1.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyppeteer==0.2.6
pyquery==1.4.3
pyRFC3339==1.0
pyrsistent==0.18.0
PySocks==1.7.1
python-apt==1.6.5+ubuntu0.5
python-crontab==2.4.2
python-dateutil==2.8.2
python-debian==0.1.32
python-slugify==6.1.2
python-socks==2.0.3
python3-gearman==0.1.0
pyttsx3==2.90
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.25
PyYAML==3.12
pyzmq==25.0.0
queuelib==1.6.2
regex==2022.3.2
reportlab==3.4.0
requests==2.23.0
requests-file==1.5.1
requests-html==0.10.0
requests-unixsocket==0.1.5
roman==2.0.0
schedule==1.1.0
Scrapy==2.6.2
SecretStorage==2.3.1
selenium==3.141.0
semantic-version==2.8.5
Send2Trash==1.8.0
service-identity==21.1.0
setuptools-rust==1.1.2
simplejson==3.13.2
six==1.16.0
smmap==5.0.0
smmap2==3.0.1
sniffio==1.2.0
snscrape==0.3.4
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.2.4
starlette==0.19.1
system-service==0.3
systemd-python==234
terminado==0.12.1
testpath==0.6.0
text-unidecode==1.3
tldextract==3.1.2
tomli==1.2.3
tornado==6.1
tqdm==4.64.0
traitlets==4.3.3
twint==2.1.20
Twisted==22.4.0
twitter-scraper==0.4.4
typing_extensions==4.1.1
tzdata==2022.2
tzlocal==4.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.11
usb-creator==0.3.3
uvicorn==0.17.0
virtualenv==15.1.0
w3lib==1.22.0
wadllib==1.3.2
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.3.1
websockets==9.1
Werkzeug==2.0.3
WTForms==3.0.0
xkit==0.0.0
yarl==1.7.2
yfinance==0.1.70
youtube-dl==2021.12.17
zipp==3.6.0
zope.interface==5.4.0
